+++ plugins/vsplit.ts
@@ -1,10 +1,4 @@
 import { defineNuxtPlugin } from 'nuxt/app'
 
-// Universal fallback registration so resolveDirective('split') always succeeds
-export default defineNuxtPlugin((nuxtApp) => {
-  nuxtApp.vueApp.directive('split', {
-    getSSRProps() {
-      return {}
-    }
-  })
-})
+// No-op: replaced by split-type.client.ts and split-type.server.ts
+export default defineNuxtPlugin(() => {})